+++ apps/prairielearn/src/server.js
@@ -1002,7 +1002,7 @@ module.exports.initExpress = function () {
     '/pl/course_instance/:course_instance_id/instructor/instance_question/:instance_question_id/clientFilesQuestion',
     [
       require('./middlewares/selectAndAuthzInstanceQuestion'),
-      require('./pages/clientFilesQuestion/clientFilesQuestion'),
+      require('./pages/clientFilesQuestion/clientFilesQuestion')(),
     ],
   );
 
@@ -1010,7 +1010,7 @@ module.exports.initExpress = function () {
     '/pl/course_instance/:course_instance_id/instructor/instance_question/:instance_question_id/generatedFilesQuestion',
     [
       require('./middlewares/selectAndAuthzInstanceQuestion'),
-      require('./pages/generatedFilesQuestion/generatedFilesQuestion'),
+      require('./pages/generatedFilesQuestion/generatedFilesQuestion')(),
     ],
   );
 
@@ -1034,7 +1034,7 @@ module.exports.initExpress = function () {
     '/pl/course_instance/:course_instance_id/instructor/instance_question/:instance_question_id/submission/:submission_id/file',
     [
       require('./middlewares/selectAndAuthzInstanceQuestion'),
-      require('./pages/submissionFile/submissionFile'),
+      require('./pages/submissionFile/submissionFile')(),
     ],
   );
 
@@ -1362,7 +1362,7 @@ module.exports.initExpress = function () {
     '/pl/course_instance/:course_instance_id/instructor/question/:question_id/clientFilesQuestion',
     [
       require('./middlewares/selectAndAuthzInstructorQuestion'),
-      require('./pages/clientFilesQuestion/clientFilesQuestion'),
+      require('./pages/clientFilesQuestion/clientFilesQuestion')(),
     ],
   );
 
@@ -1371,7 +1371,7 @@ module.exports.initExpress = function () {
     '/pl/course_instance/:course_instance_id/instructor/question/:question_id/generatedFilesQuestion',
     [
       require('./middlewares/selectAndAuthzInstructorQuestion'),
-      require('./pages/generatedFilesQuestion/generatedFilesQuestion'),
+      require('./pages/generatedFilesQuestion/generatedFilesQuestion')(),
     ],
   );
 
@@ -1380,7 +1380,7 @@ module.exports.initExpress = function () {
     '/pl/course_instance/:course_instance_id/instructor/question/:question_id/submission/:submission_id/file',
     [
       require('./middlewares/selectAndAuthzInstructorQuestion'),
-      require('./pages/submissionFile/submissionFile'),
+      require('./pages/submissionFile/submissionFile')(),
     ],
   );
 
@@ -1520,19 +1520,19 @@ module.exports.initExpress = function () {
   );
   app.use(
     '/pl/course_instance/:course_instance_id/instance_question/:instance_question_id/clientFilesQuestion',
-    require('./pages/clientFilesQuestion/clientFilesQuestion'),
+    require('./pages/clientFilesQuestion/clientFilesQuestion')(),
   );
 
   // generatedFiles
   app.use(
     '/pl/course_instance/:course_instance_id/instance_question/:instance_question_id/generatedFilesQuestion',
-    require('./pages/generatedFilesQuestion/generatedFilesQuestion'),
+    require('./pages/generatedFilesQuestion/generatedFilesQuestion')(),
   );
 
   // Submission files
   app.use(
     '/pl/course_instance/:course_instance_id/instance_question/:instance_question_id/submission/:submission_id/file',
-    require('./pages/submissionFile/submissionFile'),
+    require('./pages/submissionFile/submissionFile')(),
   );
 
   // legacy client file paths
@@ -1767,19 +1767,19 @@ module.exports.initExpress = function () {
   );
   app.use('/pl/course/:course_id/question/:question_id/clientFilesQuestion', [
     require('./middlewares/selectAndAuthzInstructorQuestion'),
-    require('./pages/clientFilesQuestion/clientFilesQuestion'),
+    require('./pages/clientFilesQuestion/clientFilesQuestion')(),
   ]);
 
   // generatedFiles
   app.use('/pl/course/:course_id/question/:question_id/generatedFilesQuestion', [
     require('./middlewares/selectAndAuthzInstructorQuestion'),
-    require('./pages/generatedFilesQuestion/generatedFilesQuestion'),
+    require('./pages/generatedFilesQuestion/generatedFilesQuestion')(),
   ]);
 
   // Submission files
   app.use('/pl/course/:course_id/question/:question_id/submission/:submission_id/file', [
     require('./middlewares/selectAndAuthzInstructorQuestion'),
-    require('./pages/submissionFile/submissionFile'),
+    require('./pages/submissionFile/submissionFile')(),
   ]);
 
   // legacy client file paths
@@ -1829,6 +1829,23 @@ module.exports.initExpress = function () {
       next();
     },
     require('./pages/publicQuestions/publicQuestions'),
+  ]);
+
+  // Client files for questions
+  app.use(
+    '/pl/public/course/:course_id/question/:question_id/clientFilesQuestion',
+    require('./pages/clientFilesQuestion/clientFilesQuestion')({ publicEndpoint: true }),
+  );
+
+  // generatedFiles
+  app.use(
+    '/pl/public/course/:course_id/question/:question_id/generatedFilesQuestion',
+    require('./pages/generatedFilesQuestion/generatedFilesQuestion')({ publicEndpoint: true }),
+  );
+
+  // Submission files
+  app.use('/pl/public/course/:course_id/question/:question_id/submission/:submission_id/file', [
+    require('./pages/submissionFile/submissionFile')({ publicEndpoint: true }),
   ]);
 
   //////////////////////////////////////////////////////////////////////