+++ packages/html-ejs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prairielearn/html-ejs
 
+## 1.1.14
+
+### Patch Changes
+
+- 2da23ab: Upgrade all JavaScript dependencies
+- Updated dependencies [2da23ab]
+  - @prairielearn/html@3.1.2
+
 ## 1.1.13
 
 ### Patch Changes