+++ packages/postgres-tools/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prairielearn/postgres-tools
 
+## 1.1.12
+
+### Patch Changes
+
+- 2da23ab: Upgrade all JavaScript dependencies
+- Updated dependencies [2da23ab]
+  - @prairielearn/postgres@1.8.1
+
 ## 1.1.11
 
 ### Patch Changes